requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.27.2
pyarrow==17.0.0
pandas==2.2.3
